--- app/src/main/java/com/example/myapplicationlvl1/SimpleTextWatcher.kt
@@ -1,15 +1,0 @@
-package com.example.myapplicationlvl1
-
-import android.text.Editable
-import android.text.TextWatcher
-
-abstract class SimpleTextWatcher : TextWatcher {
-    override fun beforeTextChanged(s: CharSequence?, start: Int, count: Int, after: Int) {
-    }
-
-    override fun onTextChanged(s: CharSequence?, start: Int, count: Int, after: Int) {
-    }
-
-    override fun afterTextChanged(p0: Editable?) {
-    }
-}
