+++ app/src/main/java/com/example/myapplicationlvl1/data/storage/DataSource.kt
@@ -1,4 +1,4 @@
-package com.example.myapplicationlvl1
+package com.example.myapplicationlvl1.data.storage
 
 interface DataSource {
