+++ app/src/main/java/com/example/myapplicationlvl1/utils/Constants.kt
@@ -1,0 +1,6 @@
+package com.example.myapplicationlvl1.utils
+
+object Constants {
+    const val LOGIN_KEY = "LOGIN_KEY"
+    const val PASSWORD_KEY = "PASSWORD_KEY"
+}
