+++ app/src/main/java/com/example/myapplicationlvl1/utils/extensions/CapitalizeWords.kt
@@ -1,0 +1,4 @@
+package com.example.myapplicationlvl1.utils.extensions
+
+fun String.capitalizeWords() = split(" ")
+    .joinToString(" ") { it -> it.lowercase().replaceFirstChar { it.uppercase() } }
