+++ app/src/main/res/layout/activity_register.xml
@@ -35,7 +35,7 @@
         app:layout_constraintTop_toBottomOf="@id/textViewHello" />
 
     <com.google.android.material.textfield.TextInputLayout
-        android:id="@+id/text_input_layout_email"
+        android:id="@+id/textInputLayoutEmail"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:layout_marginTop="20dp"
